+++ src/__tests__/fileMock.ts
@@ -1,6 +1,5 @@
 /**
- * This file is use to mock asset imports in TSX files, mostly
- * used in our case to mock SVG files.
+ * This file is used to mock asset imports in TSX files.
  */
 
 export default "div";